fixes fetch issues with current_user

+++ react/src/components/CommentTile.js
@@ -7,11 +7,13 @@ class CommentTile extends Component {
 
   render() {
     let editDeleteOptions = null
-    if(this.props.comment.user_id==this.props.current_user.id) {
-      editDeleteOptions = <p>
-        <a href={`/comments/` + this.props.comment.id + `/edit`}>Edit Comment</a>
-         <a onClick={() => this.props.handleDelete(this.props.comment)}>Delete Comment</a>
-      </p>
+    if(this.props.current_user) {
+      if(this.props.comment.user_id==this.props.current_user.id) {
+        editDeleteOptions = <p>
+          <a href={`/comments/` + this.props.comment.id + `/edit`}>Edit Comment</a>
+           <a onClick={() => this.props.handleDelete(this.props.comment)}>Delete Comment</a>
+        </p>
+      }
     }
 
     return(
